Refresh page more efficiently.
When drawing window state we should not call `wrefresh()` for every
window as that causes the virtual screen to be written through to the
user immediately. Instead, we should only do this once the entire
screen has been drawn. So we call `wnoutrefresh()` for every window
and `doupdate()` once every window has been written.