Get rid of hidden pane; use search pane in its place.
We can absorb what the hidden pane was doing into the functionality
of the search pane and have keypresses go through its keypress
callback function. As part of this, move the search pane back into
wadgadget.c as it needs to be tightly coupled to the other code here.
This will help with implementing the search functionality, anyway.