Rollback optimization when we write new directory repeatedly.
Every time we change something we must write a new copy of the WAD
directory. However, the directory is always at the end of the WAD file.
Therefore it is possible to overwrite it instead of repeatedly writing
new copies. The only issue is that we don't want to ever have the file
be in an inconsistent state; we can avoid this by temporarily rolling
back to the previous WAD directory if it is still in the file. This
way, if something goes wrong while writing the new lump/directory, we
at least still have a well-formed WAD.