Add auto-expanding list of games.
A cute feature reminiscent of NWT, which listed all the games it supported
on startup. By NWT 1.4, the list had grown quite long, and yet it was
before even Final Doom was released. Since we support arbitrary window
sizes we can automatically adjust the number of games that are shown
depending on the screen width. There's a nice long list of games that
use the WAD file format.