Add additional optimization when rewriting lump.
If we rewrite the same lump multiple times, the last version of the
lump will be at the EOF. We can exploit this and overwrite the old copy,
avoiding filling the WAD with many copies of the same lump. This should
be useful eg. when editing texture lumps as it will allow small,
incremental changes without using a ton of space.