Further improvements when running subcommand.
Factor out fork/exec to a separate function, and reuse the form of the
Windows _spawnv() API since it's what we really want and maybe this will
get ported to Windows in the future. Suspend curses while running the
subcommand and print a message indicating that the file is being opened
(since otherwise it can seem like the program froze). Finally, add an
ifdef for macOS where we can use the `open` command instead.