Add special handling for SIGTSTP (^Z)
While waiting for an editor program, it is possible that the user
does not want to wait (eg. if we launched the Gimp, the user changed
and saved the file, but wants to keep the program running). So catch
the ^Z signal and when it occurs, leave the program running in the
background.