Add contexts system to allow concurrent file access.
This allows the same code to be used to read and write from different
locations of the file depending on the context, which can be changed
using `vfswitchcontext()`. The approach here performs the seek lazily,
just before the file operation occurs. This allows multiple file ops
on the same context with seeking only taking place if and when it is
needed.