Move "make WAD" binding back to F9.
Instead of using an F-key, we can use ^G for "mark pattern". This will
allow us to free up F4, which is the standard Orthodox File Manager
binding for "Edit". There's an additional benefit that "Export WAD"
now comes after "Export" which means we can just call it that instead
of "Export to files".