Rename W_WriteDirectory() to W_CommitChanges().
I believe this is a clearer name for the function, as it makes the way
the API has to be used more clear. Less about the fine details ("the
directory needs to be written") and more about the fact that any
changes made to the file need a call to this function to take effect.