Rework delete action.
We can use `VFS_Refresh()` to jump to an appropriate location here too.
But for this to work we need to change `VFS_Remove()` to not change the
directory contents (matching others like `VFS_Rename()` that also do
not change). While we're here, we can also change the action code to
use `VFS_IterateSet()` which is more efficient.