Don't open the same directory twice.
If the directory is already open, we can just add a reference and
return the same directory. This solves several corner cases: one is that
if the same directory is open in both panes, a change in one updates
both simultaneously. Secondly, the Compact WAD action becomes a lot
simpler and more obviously correct: there is no need for any special
handling in the case where the user compacts a WAD that is open in the
other pane.