Fix bug when importing both PNAMES and TEXTURE1.
Alphabetical order ensures we will import PNAMES before TEXTURE1 (as
desired) but the texture import code was not finding the new lump as it
was using the VFS interface, and the VFS directory structure is only
updated when the update API is called. Instead, perform the lookup via
the WAD API.  As part of this we can also eliminate some duplication in
the bundle code.