Save UDS pager so it can be reopened.
If the user is referring to docs to interpret a hexdump, it is helpful
to be able to switch between the two. Forcing the user to navigate all
over again to see the text they were just looking at is not great.

Part of this is a refactor of P_RunPager so that it does not instantiate
a pager itself. That is now the job of the caller.