Add record grouping to hexdump.
A lot of lumps consist of an array of fixed-length records. It's nice
to be able to see these arranged in a readable way. Identify these
lumps by name and bold the first byte of each record, and set the
column count to be a factor or multiple of the record length.