Allow multiple stacks on one screen.
Some quite a few kinks to iron out, and the behavior needs to be tweaked
(not all pagers should split the screen). But the core behavior is all
here.

We extend the "current stack" to support multiple stacks, with an
"active stack" that is the current one accepting input. To display the
multiple stacks on screen, we need to assign them to different parts
of the screen; the recalculate function does this by setting the
top and bottom line fields.