More rearranging of key combos.
Ctrl-C makes more intuitive sense as "Copy", since it's usually used
for copy/paste (even though in this case it's a different kind of
action). Ctrl-E also works better for "Rename", since (1) the 'E' is a
character in the word "rEname", and (2) Win-E is the key combo that
Windows Explorer uses for rename.

To free up 'E' we do need to move other commands off that key. The
Shell command can use Ctrl-O which works since 'O' is in the word
"cOmmand". In the texture editor we can move Ctrl-E to Ctrl-F ("edit as
File") since the "make WAD file" command doesn't exist there. Finally
we can just get rid of the still-unimplemented "Edit" action, which was
only present in the WAD view, and it's unclear how it was going to be
different from just pressing enter to view/edit.