Always choose an appropriate transparent color.
Usually we try to use the "traditional" Doom transparency index, #247,
but there are some rare cases (notably in doom2.wad's TITLEPIC) where
that color is actually used as part of the image. So we always make
sure to check that the transparency index isn't used in the actual
picture data, and if it is, we pick a different color.

Fixes #25.