Fix handling of tables of links
As found with the previous change to support clicking links,
"byte offset from start of line" (the `.offset` field) is not the same
as screen column. So change the keyboard navigation to use the new
`LinkColumnRange` function and fix this too. The `.offset` field now
becomes a value that is entirely opaque to the main pager code.