Add `activate_link` callback to open pager links
By doing this as a "method" rather than by hooking an action, we will
be able to reuse the method for mouse clicks too.

I don't much like the function signature for `activate_link` though. The
separation of concerns between `pager` and `pager_config` has become
kind of a mess.