Send scroll keys, not cursor keys
A subtle distinction, but if we scroll up, we just want to move the
window up, not move a cursor within that window (that in the case of the
pager for example has a secondary effect of sometimes scrolling the
window up). This should also avoid any unexpected behaviors with code
that expects `KEY_UP` etc. to be literal keyboard presses.