Replace `file_type` enum with a struct
This allows us to decouple the filesystem code from some of its
implementations - the `vfs.h` header no longer refers to textures,
pnames or palettes. We do need to now add some new includes that I'm not
very happy about but this is a first step before we can replace those
too.