Make Escape "exit" to parent directory
We still keep Escape to quit the program, but it only has that effect
once we reach the level of a real filesystem directory; all other types
walk up one level to the parent. This means that when, for example,
editing textures, we exit to the WAD and then back to its containing
directory before quitting the program. I think this is a far less
surprising behavior than the previous behavior of immediately quitting
the program.

This is particularly appropriate in the case of the palette database,
which we only ever expect to open temporarily; having the exit key take
the user back to the previous screen makes a lot of sense.