Use filename for serial numbers, not inode#
We were previously using inode number, and there's an elegance in this,
but there are some problems with this:

* It does not properly handle hardlinks, which can make it possible for
  there to be two files in the same directory with the same inode
  number.
* Not all operating systems / filesystems have a concept of inode
  numbers. In particular, Windows does not have inode numbers and the
  closest equivalent (NTFS file IDs) appears to only be accessible if
  you can open the file.

Instead, let's just do it ourselves and use filenames. To do this, we
keep a sorted list of strings that we refer to when refreshing the
directory. The memory address of the string can then be used as the
serial number to give a unique number.