Extend keypress callback to return bool
This will allow panes to indicate whether they handled a particular
keypress or whether it should be passed on to the next pane in the
stack. In the majority of cases they should return true, but there are
cases like the browser interface where it's useful to have multiple
visible panes that each watch for keypresses independently (ie. the
search pane and directory panes).