Stop building macOS binaries on x86
As per actions/runner-images#13046, the `macos-13` runner is now
discontinued, and this has caused the CI to break. The website says that
`macos-15-intel` "is currently planned as the last supported image based
on Intel CPU". Rather than wait around for this to break, let's just
kill the x86 build.